Move add tracks into toolbar and styling fixes
