+++ src/containers/Playground.jsx
@@ -1,39 +1,19 @@
-import React, { useState } from "react";
-import ButtonPrimary from "../components/ButtonPrimary";
-import { MdOutlineAddCircleOutline } from "react-icons/md";
+import React from "react";
 import TrackContainer from "./TrackContainer";
 
 const Playground = ({
+  handleDelete,
+  tracks,
   isPlayingAll,
   mapping,
   octave,
   rootNote,
   scaleNotes,
 }) => {
-  const defaultTracks = [
-    {
-      key: 0,
-      id: 0,
-      text: "Hello world; welcome to WordPlay!",
-      octave: 2,
-      instrument: "Synth",
-      speed: 1,
-    },
-  ];
-  const [tracks, setTracks] = useState(defaultTracks);
-
-  const handleAdd = () => {
-    setTracks([...tracks, { key: tracks.length, id: Date.now(), text: "" }]);
-  };
-
-  const handleDelete = (trackId) => {
-    setTracks((prev) => prev.filter((track) => track.id !== trackId));
-  };
-
   return (
-    <div className="relative top-0 z-0 flex h-full flex-grow flex-col items-center justify-center bg-gradient-to-br from-white to-amber-50 p-10">
-      <div className="flex w-full flex-wrap items-center justify-center">
-        <div className="grid h-full w-full grid-cols-1 overflow-auto">
+    <div className="relative top-0 z-0 flex h-full flex-grow flex-col items-center justify-center bg-gradient-to-br from-white to-amber-50">
+      <div className="flex h-full w-full flex-wrap items-center justify-center">
+        <div className="grid h-full w-full grid-cols-1 overflow-auto p-10">
           {tracks.map((track) => (
             <TrackContainer
               key={track.key}
@@ -52,12 +32,6 @@ const Playground = ({
           ))}
         </div>
       </div>
-      <ButtonPrimary
-        icon={<MdOutlineAddCircleOutline className="mr-2" size="25px" />}
-        edit="w-30 mb-10 right-10 flex justify-center items-center text-center"
-        text="Add Track"
-        handleClick={handleAdd}
-      />
     </div>
   );
 };
