+++ src/containers/Toolbar.jsx
@@ -1,6 +1,7 @@
 import React from "react";
 import { BiEdit } from "react-icons/bi";
 import ButtonPrimary from "../components/ButtonPrimary";
+import { MdOutlineAddCircleOutline } from "react-icons/md";
 import * as Tone from "tone";
 
 function Toolbar({
@@ -8,6 +9,7 @@ function Toolbar({
   handlePlayAll,
   handleStopAll,
   handleHelp,
+  handleAdd,
 }) {
   Tone.Transport.bpm.value = 90; // Default to 90 bpm
   const handleBpmChange = ({ target }) => {
@@ -24,6 +26,12 @@ function Toolbar({
         handleClick={handleHelp}
       />
       <div className="flex gap-4">
+        <ButtonPrimary
+          icon={<MdOutlineAddCircleOutline className="mr-2" size="25px" />}
+          edit="w-30 right-10 flex justify-center items-center text-center"
+          text="Add Track"
+          handleClick={handleAdd}
+        />
         <ButtonPrimary text="Play All" handleClick={handlePlayAll} />
         <ButtonPrimary text="Stop All" handleClick={handleStopAll} />
         <h2 className="mt-1 text-lg font-semibold">BPM</h2>
